requirements.txt
# Core dependencies
openai>=2.3.0
python-dotenv>=1.1.0
pydantic>=2.12.0

# Testing dependencies
pytest>=8.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Code quality dependencies
pylint>=3.0.0
pytest-pylint>=0.21.0
coverage>=7.0.0

# Additional dependencies (auto-installed with above)
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
jiter==0.11.0
packaging<25,>=20
pluggy==1.6.0
pydantic_core==2.41.1
Pygments==2.19.2
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0

# Mood-to-Music Recommender dependencies
streamlit==1.39.0
spotipy==2.24.0
